Rework subgraph into modules and new types
