+++ packages/graph/hash_graph/lib/graph/src/shared/subgraph/mod.rs
@@ -1,321 +1,19 @@
-use std::{
-    collections::{
-        hash_map::{IntoIter, RawEntryMut},
-        HashMap, HashSet,
-    },
-    fmt::{Debug, Formatter},
-};
-
-use serde::{Deserialize, Serialize};
-use type_system::{
-    uri::{BaseUri, VersionedUri},
-    DataType, EntityType, PropertyType,
-};
-use utoipa::{openapi, ToSchema};
-
-use crate::{
-    knowledge::{Entity, PersistedEntity},
-    ontology::{PersistedDataType, PersistedEntityType, PersistedPropertyType},
-    shared::identifier::GraphElementIdentifier,
-    store::query::{Filter, QueryRecord},
-};
-
 mod depths;
+mod edges;
+mod query;
+mod vertices;
 
-pub use depths::*;
+use serde::Serialize;
+use utoipa::ToSchema;
 
-use crate::identifier::{
-    EntityIdentifier, EntityVersion, GraphElementEditionIdentifier, OntologyTypeVersion, Timestamp,
-};
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[serde(tag = "kind", content = "inner")]
-pub enum Vertex {
-    DataType(PersistedDataType),
-    PropertyType(PersistedPropertyType),
-    EntityType(PersistedEntityType),
-    Entity(PersistedEntity),
-}
-
-// WARNING: This MUST be kept up to date with the enum names and serde attribute, as utoipa does
-// not currently support adjacently tagged enums so we must roll our own:
-// https://github.com/juhaku/utoipa/issues/219
-impl ToSchema for Vertex {
-    fn schema() -> openapi::Schema {
-        let mut builder =
-            openapi::OneOfBuilder::new().discriminator(Some(openapi::Discriminator::new("kind")));
-
-        for (kind, schema) in [
-            ("dataType", PersistedDataType::schema()),
-            ("propertyType", PersistedPropertyType::schema()),
-            ("entityType", PersistedEntityType::schema()),
-            ("entity", PersistedEntity::schema()),
-        ] {
-            builder = builder.item(
-                openapi::ObjectBuilder::new()
-                    .property(
-                        "kind",
-                        // Apparently OpenAPI doesn't support const values, the best you can do is
-                        // an enum with one option
-                        openapi::Schema::from(
-                            openapi::ObjectBuilder::new().enum_values(Some([kind])),
-                        ),
-                    )
-                    .required("kind")
-                    .property("inner", schema)
-                    .required("inner"),
-            );
-        }
-
-        builder.into()
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, ToSchema)]
-#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
-pub enum EdgeKind {
-    /// An [`OntologyType`] can inherit from another [`OntologyType`]
-    InheritsFrom,
-    /// This [`Entity`] has an outgoing [`Link`] [`Entity`]
-    HasLink,
-    /// This [`Link`] [`Entity`] has another [`Entity`] at its end
-    HasEndpoint,
-    /// An [`Entity`] is of an [`EntityType`]
-    IsOfType,
-    /// A [`PropertyType`] or [`DataType`] can reference a [`DataType`] to constrain values
-    ConstrainsValuesOn,
-    /// An [`EntityType`] or [`PropertyType`] can reference a [`PropertyType`] to constrain
-    /// properties
-    ConstrainsPropertiesOn,
-    /// An [`EntityType`] can reference a [`Link`] [`EntityType`] to constrain the existence of
-    /// certain kinds of [`Link`]s
-    ConstrainsLinksOn,
-    /// An [`EntityType`] can reference an [`EntityType`] to constrain the target entities of
-    /// certain kinds of [`Link`]s
-    ConstrainsLinkDestinationsOn,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct OutwardEdge {
-    pub edge_kind: EdgeKind,
-    /// TODO: DOC - intended to capture reverse links too
-    pub reversed_direction: bool,
-    pub right_element: GraphElementIdentifier,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Default, Serialize)]
-pub struct Edges(HashMap<GraphElementIdentifier, HashSet<OutwardEdge>>);
-
-impl Edges {
-    #[must_use]
-    pub fn new() -> Self {
-        Self(HashMap::new())
-    }
-
-    pub fn insert(&mut self, identifier: GraphElementIdentifier, edge: OutwardEdge) -> bool {
-        match self.0.raw_entry_mut().from_key(&identifier) {
-            RawEntryMut::Vacant(entry) => {
-                entry.insert(identifier, HashSet::from([edge]));
-                true
-            }
-            RawEntryMut::Occupied(entry) => {
-                let set = entry.into_mut();
-                set.insert(edge)
-            }
-        }
-    }
-}
-
-// Necessary because utoipa can't handle HashSet
-impl ToSchema for Edges {
-    fn schema() -> openapi::Schema {
-        openapi::ObjectBuilder::new()
-            .additional_properties(Some(openapi::schema::Array::new(OutwardEdge::schema())))
-            .into()
-    }
-}
-
-impl IntoIterator for Edges {
-    type IntoIter = IntoIter<GraphElementIdentifier, HashSet<OutwardEdge>>;
-    type Item = (GraphElementIdentifier, HashSet<OutwardEdge>);
-
-    fn into_iter(self) -> Self::IntoIter {
-        self.0.into_iter()
-    }
-}
-
-impl Extend<(GraphElementIdentifier, HashSet<OutwardEdge>)> for Edges {
-    fn extend<T: IntoIterator<Item = (GraphElementIdentifier, HashSet<OutwardEdge>)>>(
-        &mut self,
-        other: T,
-    ) {
-        self.0.extend(other);
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[serde(tag = "kind", content = "inner")]
-pub enum OntologyVertex {
-    DataType(PersistedDataType),
-    PropertyType(PersistedPropertyType),
-    EntityType(PersistedEntityType),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
-#[serde(rename_all = "camelCase")]
-#[serde(tag = "kind", content = "inner")]
-pub enum KnowledgeGraphVertex {
-    Entity(PersistedEntity),
-}
-
-pub struct OntologyVertices(HashMap<BaseUri, HashMap<OntologyTypeVersion, OntologyVertex>>);
-pub struct KnowledgeGraphVertices(
-    HashMap<EntityIdentifier, HashMap<EntityVersion, KnowledgeGraphVertex>>,
-);
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, ToSchema)]
-#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
-pub enum OntologyEdgeKind {
-    /// An [`OntologyType`] can inherit from another [`OntologyType`]
-    InheritsFrom,
-    /// A [`PropertyType`] or [`DataType`] can reference a [`DataType`] to constrain values
-    ConstrainsValuesOn,
-    /// An [`EntityType`] or [`PropertyType`] can reference a [`PropertyType`] to constrain
-    /// properties
-    ConstrainsPropertiesOn,
-    /// An [`EntityType`] can reference a [`Link`] [`EntityType`] to constrain the existence of
-    /// certain kinds of [`Link`]s
-    ConstrainsLinksOn,
-    /// An [`EntityType`] can reference an [`EntityType`] to constrain the target entities of
-    /// certain kinds of [`Link`]s
-    ConstrainsLinkDestinationsOn,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, ToSchema)]
-#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
-pub enum KnowledgeGraphEdgeKind {
-    /// This [`Entity`] has an outgoing [`Link`] [`Entity`]
-    HasLink,
-    /// This [`Link`] [`Entity`] has another [`Entity`] at its end
-    HasEndpoint,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, ToSchema)]
-#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
-pub enum SharedEdgeKind {
-    /// An [`Entity`] is of an [`EntityType`]
-    IsOfType,
-}
-
-pub enum NewEdgeKind {
-    Ontology(OntologyEdgeKind),
-    KnowledgeGraph(KnowledgeGraphEdgeKind),
-    Shared(SharedEdgeKind),
-}
-
-pub struct NewOutwardEdge<E, V> {
-    kind: E,
-    /// If true, interpret this as a reversed mapping and the endpoint as the source, that is,
-    /// instead of Source-Edge-Target, interpret it as Target-Edge-Source
-    reversed: bool,
-    endpoint: V,
-}
-
-pub enum OntologyOutwardEdges {
-    ToOntology(NewOutwardEdge<OntologyEdgeKind, VersionedUri>),
-    ToKnowledgeGraph(NewOutwardEdge<SharedEdgeKind, EntityIdentifier>),
-}
-
-pub struct OntologyRootedEdges(
-    HashMap<BaseUri, HashMap<OntologyTypeVersion, Vec<OntologyOutwardEdges>>>,
-);
-
-pub struct EntityAndTimestamp {
-    entity_id: EntityIdentifier,
-    timestamp: Timestamp,
-}
-
-pub enum KnowledgeGraphOutwardEdges {
-    ToKnowledgeGraph(NewOutwardEdge<KnowledgeGraphEdgeKind, EntityAndTimestamp>),
-    ToOntology(NewOutwardEdge<SharedEdgeKind, VersionedUri>),
-}
-
-pub struct KnowledgeGraphRootedEdges(
-    HashMap<EntityIdentifier, HashMap<Timestamp, Vec<KnowledgeGraphOutwardEdges>>>,
-);
-
-pub enum NewEdges {
-    OntologyEdges(OntologyRootedEdges),
-    KnowledgeGraphEdges(KnowledgeGraphRootedEdges),
-}
-
-pub struct NewSubgraph {
-    pub roots: Vec<GraphElementEditionIdentifier>,
-    pub vertices: HashMap<GraphElementEditionIdentifier, Vertex>,
-    pub edges: NewEdges,
-    pub depths: GraphResolveDepths,
-}
+pub use self::{depths::*, edges::*, query::*, vertices::*};
+use crate::identifier::GraphElementIdentifier;
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, ToSchema)]
 #[serde(rename_all = "camelCase")]
 pub struct Subgraph {
     pub roots: Vec<GraphElementIdentifier>,
-    pub vertices: HashMap<GraphElementIdentifier, Vertex>,
+    pub vertices: Vertices,
     pub edges: Edges,
     pub depths: GraphResolveDepths,
-}
-
-impl Subgraph {
-    #[must_use]
-    pub fn new(depths: GraphResolveDepths) -> Self {
-        Self {
-            roots: Vec::new(),
-            vertices: HashMap::new(),
-            edges: Edges::new(),
-            depths,
-        }
-    }
-}
-
-impl Extend<Self> for Subgraph {
-    fn extend<T: IntoIterator<Item = Self>>(&mut self, iter: T) {
-        for subgraph in iter {
-            self.roots.extend(subgraph.roots.into_iter());
-            self.vertices.extend(subgraph.vertices.into_iter());
-            self.edges.extend(subgraph.edges.into_iter());
-        }
-    }
-}
-
-/// A [`Filter`] to query the datastore, recursively resolving according to the
-/// [`GraphResolveDepths`].
-#[derive(Deserialize, ToSchema)]
-#[serde(deny_unknown_fields, rename_all = "camelCase")]
-#[aliases(
-    DataTypeStructuralQuery = StructuralQuery<'static, DataType>,
-    PropertyTypeStructuralQuery = StructuralQuery<'static, PropertyType>,
-    EntityTypeStructuralQuery = StructuralQuery<'static, EntityType>,
-    EntityStructuralQuery = StructuralQuery<'static, Entity>,
-)]
-pub struct StructuralQuery<'q, T: QueryRecord> {
-    #[serde(bound = "'de: 'q, T::Path<'q>: Deserialize<'de>")]
-    pub filter: Filter<'q, T>,
-    pub graph_resolve_depths: GraphResolveDepths,
-}
-
-// TODO: Derive traits when bounds are generated correctly
-//   see https://github.com/rust-lang/rust/issues/26925
-impl<'q, T> Debug for StructuralQuery<'q, T>
-where
-    T: QueryRecord<Path<'q>: Debug>,
-{
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("StructuralQuery")
-            .field("filter", &self.filter)
-            .field("graph_resolve_depths", &self.graph_resolve_depths)
-            .finish()
-    }
 }
